Implemented warm up time for abilities
abilityDuration is how long an ability is active when its triggered,
warmUpTime is how long you have to wait to activate the ability again
after it has ceased to be active.
Therefore, an ability which is provides some sort of bonus for 5
seconds and has a warmup of 5 seconds can be used every 10 seconds.